Cleaning
- fix win terminal length issue
- add nice badges
- move fixtures
