+++ README.md
@@ -1,9 +1,16 @@
 pytest_parallel
 ===============
 
-## Introduction ##
+**pytest_parallel** extends [PyTest](http://pytest.org) to support parallel testing using mpi4py.
 
-**pytest_parallel** is a PyTest plugin that allows you to implement MPI parallel tests using PyTest and mpi4py.
+[![Python 3](https://img.shields.io/static/v1?label=Python&logo=Python&color=3776AB&message=3)](https://www.python.org/)
+![CI workflow](https://github.com/onera/pytest_parallel/actions/workflows/test.yml/badge.svg)
+[![Linux OK](https://img.shields.io/static/v1?label=Linux&logo=Linux&color=yellow&message=%E2%9C%93)](https://en.wikipedia.org/wiki/Linux)
+[![macOS OK](https://img.shields.io/static/v1?label=macOS&logo=Apple&color=silver&message=%E2%9C%93)](https://en.wikipedia.org/wiki/macOS)
+[![Windows OK](https://img.shields.io/static/v1?label=Windows&logo=Windows&color=white&message=%E2%9C%93)](https://en.wikipedia.org/wiki/Windows)
+![License-MPL](https://img.shields.io/badge/license-MPL%202.0-blue.svg)
+
+## Introduction ##
 
 ```Python
 import pytest_parallel
@@ -15,7 +22,7 @@ def test_fail_one_rank(comm):
   if comm.Get_rank()==1: assert False
 ```
 
-Here we declare a test that should run on two processes. When the test suite is run, the test will execute on two MPI processes. The `comm` fixture is an mpi4py communicator that is private to the test.
+Here a test that should run on two processes is declared. When the test suite is run, the test will execute on two MPI processes. The `comm` fixture is an mpi4py communicator that is private to the test.
 
 The test can be run on two processes, e.g. with:
 
@@ -23,17 +30,18 @@ The test can be run on two processes, e.g. with:
 mpirun -np 2 pytest --color=yes test_pytest_parallel.py
 ```
 
-And this will result in the following output:
+And the following output will be produced:
 
 ![example failing test](doc/images/test_fail.png)
 
-If the test is launch with too few MPI processes then it is skipped. For instance, launching it with:
+If there is not enough MPI processes to run the test, it will be skipped.
+For instance, the following launching command:
 
 ```Bash
 mpirun -np 1 pytest --color=yes test_pytest_parallel.py
 ```
 
-would result in:
+would lead to:
 
 ![example skipped test](doc/images/test_skip.png)
 
@@ -45,7 +53,7 @@ The `comm` fixture that you get when decorating your test with `pytest_parallel.
 
 The `pytest_parallel.mark.parallel(n_procs)` decorator takes one argument, `n_procs`.
 
-`n_procs` is generally an integer that specify the size of the communicator that will be given to the test through the `comm` fixture.
+`n_procs` is generally an integer that specifies the size of the communicator that will be given to the test through the `comm` fixture.
 
 `n_procs` can also be a list of integers to parametrize the test. For instance, the following test
 
@@ -59,7 +67,7 @@ will run two times: once with `comm` being a communicator of size 2, once with `
 
 ## Schedulers ##
 
-**pytest_parallel** comes with 3 schedulers. To understand how they work, let's take the following example:
+**pytest_parallel** comes with three kind of schedulers. To understand how they work, let's take the following example:
 
 ```Python
 import pytest_parallel
@@ -111,7 +119,7 @@ We see that processes 0,1 and 2 wait at the first step for process 3 to finish.
 
 ### Dynamic scheduler ###
 
-The **dynamic** scheduler spawns a new MPI process that act as the master scheduler and sends work to the original processes. The scheduler tries to schedule tests requiring the most processes first. The scheduler tries to send work to idle process until all the processes are busy executing one test, or when not enough processes are ready accept a test. It then waits for a signal that workers have finished their test to schedule further work.
+The **dynamic** scheduler spawns a new MPI process which acts as the master scheduler and sends work to the original processes. The scheduler tries to schedule tests requiring the most processes first. The scheduler tries to send work to idle process until all the processes are busy executing one test, or when not enough processes are ready to accept a test. It then waits for a signal that workers have finished their test to schedule further work.
 
 Example:
 ![static scheduler sequence diagram - bad case](doc/images/dyn_anim.png)
