+++ pytest_mpi_check/plugin.py
@@ -1,1 +1,2 @@
 from pytest_parallel.plugin import *
+from pytest_parallel.fixtures import *
