+++ pytest_parallel/plugin.py
@@ -17,23 +17,6 @@ def pytest_addoption(parser):
         default="sequential",
     )
 
-## TODO move to fixtures.py
-@pytest.fixture
-def comm(request):
-    """
-    Only return a previous MPI Communicator (build at prepare step )
-    """
-    return request._pyfuncitem._sub_comm  # TODO clean
-
-
-## TODO backward compatibility begin
-@pytest.fixture
-def sub_comm(request):
-    return request._pyfuncitem._sub_comm
-
-
-## TODO backward compatibility end
-
 
 # --------------------------------------------------------------------------
 @pytest.hookimpl(trylast=True)
