+++ test/pytest_parallel_refs/terminal_parametrize
@@ -1,4 +1,4 @@
-============================= test session starts ==============================
+[=]+ test session starts [=]+
 platform [^\n]*
 cachedir: [^\n]*
 ?(metadata: [^\n]*)?
@@ -19,4 +19,4 @@ collecting ... [\s]*collected 8 items[\s]*
 )(?=.*test_parametrize.py::test_mark_mpi_decorator_with_list_and_other_pytest_deco\[2-4\] PASSED
 ).*
 
-============================== 8 passed in [^\n]*s ===============================
+[=]+ 8 passed in [^\n]*s [=]+
