+++ test/pytest_parallel_refs/terminal_scheduling
@@ -1,4 +1,4 @@
-============================= test session starts ==============================
+[=]+ test session starts [=]+
 platform [^\n]*
 cachedir: [^\n]*
 ?(metadata: [^\n]*)?
@@ -19,4 +19,4 @@ collecting ... [\s]*collected 8 items[\s]*
 )(?=.*test_scheduling.py::test_7\[2\] PASSED
 ).*
 
-============================== 8 passed in [^\n]*s ===============================
+[=]+ 8 passed in [^\n]*s [=]+
