+++ test/pytest_parallel_refs/terminal_seq
@@ -1,4 +1,4 @@
-============================= test session starts ==============================
+[=]+ test session starts [=]+
 platform [^\n]*
 cachedir: [^\n]*
 ?(metadata: [^\n]*)?
@@ -13,15 +13,15 @@ collecting ... [\s]*collected 2 items[\s]*
 )(?=.*test_seq.py::test_false FAILED
 ).*
 
-=================================== FAILURES ===================================
-__________________________________ test_false __________________________________
+[=]+ FAILURES [=]+=
+[_]+ test_false [_]+
 
--------------------------------- On rank 0 of 1 --------------------------------
+[-]+ On rank 0 of 1 [-]+
     def test_false\(\):
->     assert 0
-E     assert 0
+>       assert 0
+E       assert 0
 
 [^\n]*test_seq.py:6: AssertionError
-=========================== short test summary info ============================
+[=]+ short test summary info [=]+
 FAILED [^\n]*/test_seq.py::test_false[^\n]*
-========================= 1 failed, 1 passed in [^\n]*s ==========================
+[=]+ 1 failed, 1 passed in [^\n]*s [=]+
