+++ test/pytest_parallel_refs/terminal_two_fail_tests_one_proc
@@ -1,4 +1,4 @@
-============================= test session starts ==============================
+[=]+ test session starts [=]+
 platform [^\n]*
 cachedir: [^\n]*
 ?(metadata: [^\n]*)?
@@ -13,32 +13,32 @@ collecting ... [\s]*collected 2 items[\s]*
 )(?=.*test_two_fail_tests_one_proc.py::test_fail_2\[1\] FAILED
 ).*
 
-=================================== FAILURES ===================================
-(?=.*________________________________ test_fail_1\[1\] ________________________________
+[=]+ FAILURES [=]+
+(?=.*[_]+ test_fail_1\[1\] [_]+
 
--------------------------------- On rank 0 of 1 --------------------------------
+[-]+ On rank 0 of 1 [-]+
 comm = <mpi4py.MPI.Intracomm object at [^\n]*>
 
     @pytest_parallel.mark.parallel\(1\)
     def test_fail_1\(comm\):
->     assert 0
-E     assert 0
+>       assert 0
+E       assert 0
 
 [^\n]*test_two_fail_tests_one_proc.py:6: AssertionError
-)(?=.*________________________________ test_fail_2\[1\] ________________________________
+)(?=.*[_]+ test_fail_2\[1\] [_]+
 
--------------------------------- On rank 0 of 1 --------------------------------
+[-]+ On rank 0 of 1 [-]+
 comm = <mpi4py.MPI.Intracomm object at [^\n]*>
 
     @pytest_parallel.mark.parallel\(1\)
     def test_fail_2\(comm\):
->     assert 0
-E     assert 0
+>       assert 0
+E       assert 0
 
 [^\n]*test_two_fail_tests_one_proc.py:11: AssertionError
 ).*
-=========================== short test summary info ============================
+[=]+ short test summary info [=]+
 (?=.*FAILED [^\n]*test_two_fail_tests_one_proc.py::test_fail_1\[1\][^\n]*
 )(?=.*FAILED [^\n]*test_two_fail_tests_one_proc.py::test_fail_2\[1\][^\n]*
 ).*
-============================== 2 failed in [^\n]* ===============================
+[=]+ 2 failed in [^\n]* [=]+
