+++ test/pytest_parallel_refs/terminal_two_fail_tests_two_procs
@@ -1,4 +1,4 @@
-============================= test session starts ==============================
+[=]+ test session starts [=]+
 platform [^\n]*
 cachedir: [^\n]*
 ?(metadata: [^\n]*)?
@@ -13,52 +13,52 @@ collecting ... [\s]*collected 2 items[\s]*
 )(?=.*test_two_fail_tests_two_procs.py::test_fail_2\[2\] FAILED
 ).*
 
-=================================== FAILURES ===================================
-(?=.*________________________________ test_fail_1\[2\] ________________________________
+[=]+ FAILURES [=]+
+(?=.*[_]+ test_fail_1\[2\] [_]+
 
--------------------------------- On rank 0 of 2 --------------------------------
+[-]+ On rank 0 of 2 [-]+
 comm = <mpi4py.MPI.Intracomm object at [^\n]*>
 
     @pytest_parallel.mark.parallel\(2\)
     def test_fail_1\(comm\):
->     assert 0
-E     assert 0
+>       assert 0
+E       assert 0
 
 [^\n]*test_two_fail_tests_two_procs.py:6: AssertionError
 
--------------------------------- On rank 1 of 2 --------------------------------
+[-]+ On rank 1 of 2 [-]+
 comm = <mpi4py.MPI.Intracomm object at [^\n]*>
 
     @pytest_parallel.mark.parallel\(2\)
     def test_fail_1\(comm\):
->     assert 0
-E     assert 0
+>       assert 0
+E       assert 0
 
 [^\n]*test_two_fail_tests_two_procs.py:6: AssertionError
-)(?=.*________________________________ test_fail_2\[2\] ________________________________
+)(?=.*[_]+ test_fail_2\[2\] [_]+
 
--------------------------------- On rank 0 of 2 --------------------------------
+[-]+ On rank 0 of 2 [-]+
 comm = <mpi4py.MPI.Intracomm object at [^\n]*>
 
     @pytest_parallel.mark.parallel\(2\)
     def test_fail_2\(comm\):
->     assert 0
-E     assert 0
+>       assert 0
+E       assert 0
 
 [^\n]*test_two_fail_tests_two_procs.py:11: AssertionError
 
--------------------------------- On rank 1 of 2 --------------------------------
+[-]+ On rank 1 of 2 [-]+
 comm = <mpi4py.MPI.Intracomm object at [^\n]*>
 
     @pytest_parallel.mark.parallel\(2\)
     def test_fail_2\(comm\):
->     assert 0
-E     assert 0
+>       assert 0
+E       assert 0
 
 [^\n]*test_two_fail_tests_two_procs.py:11: AssertionError
 ).*
-=========================== short test summary info ============================
+[=]+ short test summary info [=]+
 (?=.*FAILED [^\n]*test_two_fail_tests_two_procs.py::test_fail_1\[2\][^\n]*
 )(?=.*FAILED [^\n]*test_two_fail_tests_two_procs.py::test_fail_2\[2\][^\n]*
 ).*
-============================== 2 failed in [^\n]*s ===============================
+[=]+ 2 failed in [^\n]*s [=]+
