+++ test/pytest_parallel_refs/terminal_two_fail_tests_two_procs_skip
@@ -1,4 +1,4 @@
-============================= test session starts ==============================
+[=]+ test session starts [=]+
 platform [^\n]*
 cachedir: [^\n]*
 ?(metadata: [^\n]*)?
@@ -9,7 +9,7 @@ collecting ... [\s]*collected 2 items[\s]*
 [^\n]*test_two_fail_tests_two_procs.py::test_fail_1\[2\] SKIPPED[^\n]*
 [^\n]*test_two_fail_tests_two_procs.py::test_fail_2\[2\] SKIPPED[^\n]*
 
-=========================== short test summary info ============================
+[=]+ short test summary info [=]+
 SKIPPED \[1\] [^\n]*test_two_fail_tests_two_procs.py:4: Not enough procs to execute: 2 required but only 1 available
 SKIPPED \[1\] [^\n]*test_two_fail_tests_two_procs.py:9: Not enough procs to execute: 2 required but only 1 available
-============================== 2 skipped in [^\n]*s ==============================
+[=]+ 2 skipped in [^\n]*s [=]+
