+++ test/pytest_parallel_refs/terminal_two_success_tests_two_procs
@@ -1,4 +1,4 @@
-============================= test session starts ==============================
+[=]+ test session starts [=]+
 platform [^\n]*
 cachedir: [^\n]*
 ?(metadata: [^\n]*)?
@@ -13,4 +13,4 @@ collecting ... [\s]*collected 2 items[\s]*
 )(?=.*test_two_success_tests_two_procs.py::test_success_2\[2\] PASSED
 ).*
 
-============================== 2 passed in [^\n]*s ===============================
+[=]+ 2 passed in [^\n]*s [=]+
